+++ Sources/KeyboardKit/Autocomplete/AutocompleteProvider.swift
@@ -8,18 +8,6 @@
 
 import Foundation
 
-/// [DEPRECATED]
-///
-/// > v8.0: This type will be removed in KeyboardKit 8.0 but
-/// is still used while the service is asynchronous.
-public typealias AutocompleteCompletion = (AutocompleteResult) -> Void
-
-/// [DEPRECATED]
-///
-/// > v8.0: This type will be removed in KeyboardKit 8.0 but
-/// is still used while the service is asynchronous.
-public typealias AutocompleteResult = Result<[AutocompleteSuggestion], Error>
-
 /**
  This protocol can be implemented by any classes that can be
  used to give autocomplete suggestions as the user types.
@@ -46,19 +34,9 @@ public protocol AutocompleteProvider: AnyObject {
 
 
     /// Get autocomplete suggestions for the provided `text`.
-    ///
-    /// > v8.0: This will be made async in KeyboardKit 8.0.
     func autocompleteSuggestions(
-        for text: String,
-        completion: @escaping Completion
-    )
-
-    
-    /// > v8.0: This will be removed in KeyboardKit 8.0.
-    typealias Completion = AutocompleteCompletion
-
-    /// > v8.0: This will be removed in KeyboardKit 8.0.
-    typealias CompletionResult = AutocompleteResult
+        for text: String
+    ) async throws -> [AutocompleteSuggestion]
 
 
     /// Whether or not the provider can ignore words.